display gen and sim status in column
