+++ js/main.js
@@ -6,6 +6,8 @@ import {
   genStatus,
   renderRangeParameters,
   parametersRange,
+  getWidthFilterContent,
+  renderGenSim,
 } from "./menu/filter/filter.js";
 import {
   mouseDown,
@@ -20,7 +22,6 @@ const canvas = document.getElementById("canvas");
 const ctx = canvas.getContext("2d");
 
 const manipulationTools = document.getElementsByClassName("manipulation-tool");
-const filterContent = document.getElementById("filter-content");
 const filter = document.getElementById("filter");
 const filters = document.getElementById("filters");
 
@@ -156,14 +157,6 @@ function hideInputModal() {
   modal.style.display = "none";
 }
 
-function getWidthFilterContent() {
-  filterContent.style.display = "flex";
-  const rangeFilters = document.getElementById("range-filters");
-  const width = rangeFilters.offsetWidth;
-  filterContent.style.display = "none";
-  return `${width}px`;
-}
-
 document.getElementById("input-file").addEventListener("change", (event) => {
   for (const file of event.target.files) {
     if (!file.name.endsWith("edm4hep.json")) {
@@ -227,8 +220,7 @@ document
     const width = getWidthFilterContent();
     filter.style.width = width;
 
-    bits.render(filters);
-    genStatus.render(filters);
+    renderGenSim(bits, genStatus, filters);
 
     const pdgToggle = new PdgToggle("show-pdg");
     pdgToggle.init(() => {
